requirements.txt
alembic==1.10.2
bcrypt==4.0.1
black==23.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
